Tentato di fare il modifica vacanza. Problema con l'aggiornamento della view prima che la vacanza sia caricata (sempre vuota). Prova ad usare LiveData su vacanza e partecipanti, magari tutto in altra activity. - in recenti, aggiungi un header e preleva le 3 liste a seconda dell'ordine temporale - in archivio, visualizza vacanze archiviate - modifica una vacanza - rimuovi tastiera quando si seleziona la data e in generale quando non serve al campo successivo - validazione dei campi della nuova vacanza

+++ app/src/main/java/com/rbiffi/vacationfriend/AppSections/VacationList/ViewModels/VacationViewModel.java
@@ -33,11 +33,11 @@ public class VacationViewModel extends AndroidViewModel {
         return allVacations;
     }
 
-    public void delete(int v) {
+    public void delete(long v) {
         repository.delete(v);
     }
 
-    public void store(int id) {
+    public void store(long id) {
         repository.store(id);
     }
 }
